Users Management: Add support to profile and i18n translation
This patch implements the support to three pre-defined user profiles:
'kimchiuser', which does not have any host access; 'virtuser', which
has host access and virtualization rights (belongs to KVM group); and
'admin', which is a user with sudo rights.

This patch also fixes the Exceptions messages, in order to use i18n,
like Kimchi.

Change-Id: I5f9da6320f387179b31406524379dad8dd84bf57
Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>
Reviewed-on: http://9.181.129.215/56
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>