Automatic rollback network configuration on failure
When the user submit wrong network configurations, it's possible the
front-end losts access to back-end server. In this case we should be
able to rollback the configurations.

This patch introduces a timout thread and a confirm_change action. After
updating the an interface configuration, the front-end should initiate a
POST request to plugins/ginger/network/interfaces/eth0/confirm_change
with empty body. If the network configuration is incorrect and front-end
losts access to kimchi server, the timeout thread rollback the
configuration. If the network configuration is correct, the front-end
should successfully call the confirm_change action, and the result is
that the server cancels the timeout and commit the configuration.

How to test:
Fistly disable NetworkManager, it conflicts with libvirt iface API:
 systemctl disable NetworkManager
 systemctl stop NetworkManager

Suppose you have eth0 and eth1, and eth0 connects to the management
network for kimchi server.

Submit a wrong configuration for eth0 and try to confirm:
 curl -u root -H "Content-Type: application/json" -H "Accept: application/json" -X PUT -d \
  '{"ipaddr": "192.168.123.88", "netmask": "255.255.255.0", "gateway": "192.168.123.1"}' \
  http://server_ip:8000/plugins/ginger/network/interfaces/eth0

 curl -u root -H "Content-Type: application/json" -H "Accept: application/json" \
  -X POST -d '' http://server_ip:8000/plugins/ginger/network/interfaces/eth0/confirm_change

After the first curl command, the connection is lost. Wait for 10 seconds,
the connection should be restored. Then the second curl calling the
confirm_change action should return an exception. Whenever we see it
raises exception, it's an indication of the incorrect network
configuration.

Submit a correct configuration and confirm for eth1:
 curl -u root -H "Content-Type: application/json" -H "Accept: application/json" -X PUT -d \
  '{"ipaddr": "192.168.122.90", "netmask": "255.255.255.0", "gateway": "192.168.123.1"}' \
  http://server_ip:8000/plugins/ginger/network/interfaces/eth1

 curl -u root -H "Content-Type: application/json" -H "Accept: application/json" \
  -X POST -d '' http://server_ip:8000/plugins/ginger/network/interfaces/eth1/confirm_change

The curl commands should run successfully.

v2:
 Make a symbol constant for the timeout value, and remove useless
 timeout.cancel() call.

Change-Id: I8c1914f742323814af512bb4e501e06d58500ecd
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://9.181.129.215/57
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>