Improve host configuration backup back-end implementation
After discussion with the front-end developer, the back-end has changed
the following places.

1. "ident" field in the archive information is renamed to "identity".

2. The "disable value" of days_ago and counts_ago is changed to -1.

3. "include" and "exclude" field in the archive creation parameters can
be "[]". Though formally we ask front-end to not send "include" at all
when it's empty, but in implementation we try to be tolerant. The
back-end treats "[]" as empty.

4. Fix a format in the i18n error string.

v2:
Discarding slicing algorithm in v1 patch is broken. V2 fixs it.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>