host backup download: Use serve_file instead of internal redirect
When using the internal redirect, the browser thinks
  plugins/ginger/backup/archives/uuid/file
is the file to download, and the downloaded file name is just "file".
This is inconvinient for the users.

It's possible to issue a header to define the real name as follow.
  Content-Disposition: attachment; filename="xxx.tar.gz"
However when using an internal redirect, the response headers set by the
"file" handler are not reflected in the final response.

This patch makes use of cherrypy.lib.static.serve_file function to
provide the data stream with the Content-Disposition header directly.

The serve_file function returns a generator that splits file data into
64K blocks. It also processes byte range request from the browser so
that it supports resuming download and multithreaded download.

v2:
  Use serve_file function instead of HTTP redirection.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>