Rework error messages reported when updating the Power FW.
The current rc/error reporting treated any rc as a
failure. update_flash may still perform the flash even
when it returns an rc. This is probably an error with the
update_flash script that needs to be addressed, but kimchi
should still not tell the user that the flash failed when
it did complete.

The Firmware Update now logs the error and tells the user where
to look to decide if any action needs to be taken.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>