Host Backup: Refine Include and Exclude Rules
When creating a backup, if the exclude is not defined, Ginger would use
the default exlude rules. However this is not expected by the user if
he/she wants to create a custom backup with custom include rules and
empty exclude rules.

This patch refines the include and exclude rules handling as follow.

Case 1, include rules are specified:
Regardless of the exlucde rules are empty or not, use the input and do
not use the default rules.

Case 2, include rules are empty, exclude rules are specified:
Use the defualt include rules, and the user defined exclude rules.

Case 3, both include and exclude rules are specified:
Obviously, use the user defined rules for include and exclude.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>