Host Backup: Make Include and Exclude Paths Configurable
Previously the include and exclude paths are hard coded in backup.py
model. In this patch we move their definition to
plugins/ginger/ginger.conf, so the user can easily tune the default
include and exclude paths.

This patch also changes the default backup paths. /var/lib/libvirt is
not in default include any longer. The files in /var/lib/libvirt are of
either large images or auto-generated xml files. They are not host
configurations.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>