Add FW commit and reject capability.
Allow a user to commit the image running on the temporary
side to the permanent side, or reject an image running
on the temporary side and overwrite it with the one
on the permanent side. Also adding the option to prohibit
an overwrite of the permanent side when updating the fw.

Notes:
    1) If commiting the temp image to the perm side, the
    system must be booted into the temp side.
    2) If rejecting the temp side image, the system must
    be booted to the perm side.

The REST call for temp commit would look like:

    curl -k -u user:password -H 'Content-type: application/json' \
    -H 'Content-length: 0' -X POST -H 'Accept: application/json' \
    https://hostame:8001/plugins/ginger/firmware/commit

This patch also uses the 'lsmcode' command instead of
pulling the current firmware level from the files
under the /proc/device-tree directory. This allows access
to all firmware levels (temp, perm, and booted), whereas
the file structure only contained the booted level.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>