Kimchi-ginger: Declaration of Independence
We hold these truths to be self-evident, that all projects are created
equal, that they are endowed by their creator with certain unalienable
rights, that among these are configuration, compilation, installation,
i18n translation and packaging.

We, therefore, the representatives of the Kimchi-ginger development
team, solemnly publish and declare, that the Kimchi-ginger project is,
and of right ought to be independent project, that it is absolved
from all allegiance to the Kimchi upstream packaging system, and that
all previous changes in upstream configure.ac Makefile.am and kimchi.spec,
are and ought to be totally dissolved, and that as independent project,
it has full power to run ./autogen.sh, make, make check, make dist, make
install, make rpm, and to do all other packaging operations which
independent project may of right do.

Note:
The new packaging system needs a git annotated tag to determine the
package version. Before the maintainer adds a tag in the ginger master
branch, we have to add one manually as follow to have the packaging work
properly.
  git tag -a kimchi-ginger-1.2.0
  # Input kimchi-ginger-1.2.0

v2:
  Rebase to latest Ginger master. Remove m4/Changelog from Makefile.am.
v3:
  Rebase and solve conflicts.
v4:
  Delete build dependency on Kimchi.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>