Fix pyflakes and pep 8 errors
In ginger.py, "Collection" is imported but not used.
In models/firmware.py, only 1 blank line is added before
"class FirmwareModel".
In models/powermanagement.py, the argument name is "profile", but later
"powerprofile" is used.

This patch fixes these problems. When submitting new patches, authors
are encouraged to run "make check-local" in advance.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>