host-admin.js: disabling OK button when processing
V1 -> V2:
1) When generating a backup report, disable all the item including input
box as well as buttons in UI.
2) Popup error when the backup is too large to operate as well as
changing the UI back to function.

Each time the OK button is pressed a new backup process is fired.
This patch makes the Ok button disable after the first time the
button is pressed to avoid this behavior. The cursor is also
changed to the 'waiting' cursor while the backup is being
generated.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>
Signed-off-by: Wen Wang <wenwang@linux.vnet.ibm.com>