Deal with bad hdd sensor data
If the last item returned from hddtemp for a disk isn't a temperature, then
no disk temperatures are returned.

Here are some examples of bad data:

/dev/sda: IBM ST300MP0064: 30°C
/dev/sdb: IBM IPR-0 5D81B600:  drive supported, but it doesn't have a temperature sensor.

More data from another system:
/dev/sda: IBM IPR-10 5D831200: S.M.A.R.T. not available

===

Another odd issue: If you use the -u option and specify 'F',  you get a temp
(though it's obviously wrong):

/dev/sda: IBM ST300MP0064: 30°C
/dev/sdb: IBM IPR-0 5D81B600:  drive supported, but it doesn't have a temperature sensor.

/dev/sda: IBM ST300MP0064: 86°F
/dev/sdc: IBM IPR-0 5D813400: 32°F

===

This patch tries to parse what should be the temperature, and if it fails, it skips
adding that disk to the list, and logs an error.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>