models/interfaces.py: removing ip/netmask on blank input
This patch fixes github #1: https://github.com/kimchi-project/ginger/issues/1

API.json now allows blank ip/netmask values, which will be interpreted
by the backend as a command to remove the current IP configuration.
Note that the ip/netmask validation still exists in the frontend
(form validation) and in the backend.

The lookup function got reworked to show the static ip/netmask
configuration in the case where there isn't an active ip/netmask
in the interface. This is more consistent than the previous
behavior, where the ip/netmask remained blank even after setting
a new static ip/netmask in the case there isn't an active address.
This happens in the case where the ethernet cable isn't plugged
in the interface, for example.

Other changes:

- iface.create(), equivalent of "ifup()", won't be called if
the interface isn't active.
- _rollback_on_failure is now being used to handle the exception
instead of 'conn.changeRollback()'.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>