Debian/Ubuntu packaging files
This patch adds new files and changes existing config files
to add autoconf/automake support for Debian/Ubuntu packaging
files. In other words, this patch adds the support to generate
a .deb package for ginger.

To test it, running in a Debian/Ubuntu system:

./autogen.sh --system
make
make deb

The ginger .deb package file should be generated in the root dir
of the plug-in.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>