Github #9, #10 and #12 fixes
Issues #9 and #10 describe errors related to the absence of the
VERSION file in the zip/tarball downloaded from Ginger webpage
in Github. This happens because Github does not include in the
compressed file any file that are in .gitignore. Excluding the
VERSION file from .gitignore fixes both issues.

Issue #12 says that the ChangeLog is missing from the zip/tarball.
This was fixed before in a previous commit, but to make it
available we need to retag Ginger to force Github to create
a new tarball.

There's also a minor issue with the previous naming used for the
Ginger tags. Naming the tags as 'ginger-<version>' forces Github
to create a tarball with name 'ginger-<tag>', thus the file name
becomes 'ginger-ginger-<version>'. Since we need to retag the
project due to these issues, the tag name will be changed to
just '<version>'. To reflect that, pkg-version was changed.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>