Translation: Change process - sort output and avoid fuzzy msgs
This patch changes the makefile system to sort the output of translation
files. It also does not output fuzzy automatic translations, avoiding
errors and loss of messages.

It also includes *.gmo in .gitignore. These changes aim to make easier
the reviews of translation patches because artifacts (pot and po files)
will be cleaner.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>