Adapt Ginger to use Wok - Webserver Originated from Kimchi
- Use WokRoot instead of Root
This avoids plugin breaking due to absence of default_page
attribute, and also makes Wok models/API available to Ginger.

- Do not use kimchi.url anymore
This makes analog changes to Ginger as done by Kimchi's commit
61ffb22a8733250dc5792528773939769072713c .

- Ginger is not using cachebust's href()

- Add ginger-help.css for help pages
Since kimchi is now a plugin, kimchi.css is not accessible
anymore. Add a copy of it as ginger-help.css.

- Install ginger .mo files to default locale dir
This is necessary to adapt Ginger to changes done by Kimchi's
commit 2e90908b8e79ff293298dfcd83b166dd06d361c3 .

- Remove version restriction from wok dependency

- Fix tests

- Allow Ginger to receive wok server options in constructor
This change is necessary to adapt Ginger to changes proposed
in kimchi-wok split V2 on kimchi side.

- Move kimchi-specific network utilities to ginger
The modules network and netinfo are moved from wok to kimchi
plugin. Fix it by adding netinfo module and network functions
used by Ginger.

Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>
Signed-off-by: Gustavo Y. Ribeiro <gyr@linux.vnet.ibm.com>