Fix user/groups id creation in user management
Some distributions, like Ubuntu, does not deliver libuser.conf, it is
possible to workaround this creating an empty /etc/libuser.conf file.
However this may create normal users/groups with system ids (< 1000).
This patch fixes this problem, always creating the users and groups
with an ID > 1000.
It also fixes problems when a group is not found.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>