Listing the information from ifcfg files
This patch adds a new control/model called Cfginterfaces and
Cfginterface to handle the configuration files of the network
interfaces. It includes changes in control and model files,
docs/API.md, i18n and Makefile.am.

A new unit test was added to support this new feature. The test
uses a python module called 'mock' that wasn't being used before,
thus an update in the README file was also made to add this package
as an unit test dependency.