Make Ginger RPM noarch
Ginger will load each module based on what is installed in the
system. Some distro-specific dependencies are not necessary,
and may make it hard to install Ginger in some cases.

Remove arch-specific dependencies, add wok and other missing
dependencies, and update README to include a openSUSE section
and to reflect those dependency changes.

Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>

Changes in V2:
- Removed python-configobj dependency (it should be in gingerbase)