Upgrade Ginger objectstore schema.
With the new structure of the objectstore changed in Wok 2.0.0, the schema of
the objectstore file needs to be checked and, if necessary, updated to reflect
the new structure.

This patch adds a step to execute an upgrade of the schema in the previous
objectstore file (older than Ginger 1.3.0) during the startup of Ginger plugin.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>