Do not rely on python-pip to install build dependencies
This patch add ginger.css file to the source control. That way the
build process does not need to rely on python-pip to install build
dependencies.

The python-pip is still required for UI development to install cython
and libsass libraries. They will be needed to generate ginger.css file
according to changes in .scss files.

When modifying the .scss files, make sure to run:

    $ make -C ui/css css

to update ginger.css accordingly.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>