Unit test for Sensors model
This unit test checks if the parsing of 'sensors -u' and
'hddtemp' is being done properly by the model.

A change in models/sensors.py was necessary to allow
the unit test to override the default temperature in the
unit test - this value is fetch from a cherrypy config that
the unit test can't override inside the body of the
lookup() function.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>