Adding CfginterfacesModel to the Capabilities API
At this moment this feature does not work for any distro
other than RHEL/Fedora and similars. The deciding factor
is the format of the configuration file, hosted in
/etc/sysconfig/network-scripts/.

Until we make the feature work for other distros, the
Capabilities API must indicate the UI (or any other caller)
that this model won't work properly in such cases.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>