Update Ginger configuration according to recent changes on Wok
The application name is restricted for the application configuration and
should never be edited. In that case, it is good to remove this kind of
information from a configuration file.

According to the recent changes on Wok, the application name will be
the plugin name with the first letter in uppercase.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>