Revert "Adding sub_nodes to __init__.py"
This patch is not necessary anymore. The real problem
was fixed in WoK. The message that appears when loading
WoK:

'Failed to import subnodes for plugin plugins.ginger.Ginger, error: Class plugins.ginger.control.sub_nodes can not be imported, error: No module named control'

is annoying but does not prevent the plug-in from being loaded.

This reverts commit 15a70c85a57ab339ed3878c6391c4a22b2d99ba4.