Github #193: Modified the ipv4 and ipv6 code to accept valid values
Modified the code for Ipv4 and Ipv6 to accept valid values and
throw proper error for invalid  values.

Added the corresponding errors in i18n.py file and modified
few errors.

Added a check in validate Ipv4 address to check whether the input
to the function is in the Ipv4 address format.

Signed-off-by: Megha Smriti <mesmriti@linux.vnet.ibm.com>