OVS bridges backend: new model/ovsbridges.py module
The module ovsbridges.py hosts all methods related to
OVS bridges handling and the implementation of OVSBridgesModel
(Collection) and OVSBridgeModel (Resource) classes.

The function run_ovsvsctl_command() is used as a broker
of run_command() to manage the case where the Openvswitch
daemon is not running. Methods that need to fire an ovs-vsctl
command do it by using run_ovsvsctl_command().

is_feature_available() will simply verify if the ovs-vsctl
command is callable. If not (return code 127) then it will
return 'False'.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>