Adding 'module' into interfaces API: sysmodule changes
The netinfo module will use the lsmod output to see if
the value retrieved by the sysfs matches lsmod output.
'lsmod' and other module functions belongs to the sysmodule
API.

This patch adds a new 'get_loaded_modules_list' to sysmodules
that can be used by netinfo. The code of this method was
extracted from the 'get_list' method of of SysModulesModel().

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>