Adding 'module' into interfaces API: model changes
In netinfo module, a new method called 'get_interfaces_kernel_module'
was created to determine the kernel module that loaded
the interface by using sysfs and matching it against the
listed of loaded modules list.

interfaces.py was changed too because the cfginterfaces
lookup() doesn't use the netinfo module, thus we need to
insert the 'module' key by hand.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>