SR-IOV for mlx5 interfaces: docs/api.md and i18n changes
The SR-IOV enablement for mlx5 interfaces is a bit different
from its antecessor, mlx4. It requires writing specific
files in the /sys filesystem for each interface loaded by
the mlx5 driver instead of simply loading specific module
parameters. It is not globally enabled.

A new generic API called 'action' was deleloped to be
used not only in this case, but any case where a specific
interface procedure/setup is required. Any available actions
will be retrieved by the lookup method of the interface.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>