Issue #249 Sorting on size gives incorrect results
v3:
Changed size to return in KBs instead of bytes for Volume Groups

v2:
Modified titles to indicate units of size

v1:
This patch fixes the issue with sorting devices
in storage tab based on size. It modifies the
data type of the size parameters returned from
corresponding APIs, to numeric.

Signed-off-by: Pooja Kulkarni <pkulkark@linux.vnet.ibm.com>