SR-IOV for ConnectX-4 pre-check: model changes
This patch implements a new function called '_mlx5_SRIOV_precheck'
in model/interfaces.py that will handle the list of pre-requirements
to enable the SR-IOV for ConnnectX-4 cards (mlx5_core driver).

To verify if the SR-IOV is enabled in the firmware, we verify
if the '/sys/class/net/<iface>/device/sriov_totalvfs' exists.
If not, we assume that SR-IOV is not enabled in the driver
and return an appropriate error. If exists, then this value
is matched against 'num_vfs' received to see if the user is
trying to set more virtual functions that the firmware allows
at that time.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>