Fixing ui/css/ginger.css license mismatch after compiling it
The way the license script works and how the ginger.css file
is generated is always causing a mismatch:

- the script checks git log to see when the file was first added
to see the start year of the license header. In ginger.css case,
2014-2016.

- since last year ginger.css is now being created by ginger.scss.
The license header created to ginger.css is the same from
ginger.scss, which is 2015-2016.

- running the license script just after compiling ginger.css from
ginger.scss will always give a mismatch because ginger.css header
isn't 2014-2016.

- if we change ginger.scss header to 2014-2016, the script will
flag ginger.scss header because this file was added in 2015.

Since ginger.scss is now the source file for ginger.css, this
patch will make ginger.css follows the same copyright year
of the scss file and put the css file in the license blacklist,
fixing this eternal mismatch.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>