SR-IOV for ConnectX-4: asynchronous backend
Tests show that we can't affort to wait for the SR-IOV
setup for ConnectX-4 to finish in the main thread. This
setup takes several seconds to complete and it's now
a 2-step process (set VF to zero, then to the desired
value), taking it even longer to complete. To solve this
problem the following changes in the backend were done
by this patch:

- SR-IOV setup is now asynchronous. A task called
'_mlx5_SRIOV_enable_task' is now launched after the
precheck from the '_mlx5_SRIOV_enable' method. To do
that it was necessary to import and use WoK TaskModel
and to use Ginger objectstore inside InterfaceModel.

- The setup is now a 2 step process. First step is to
set the virtual functions to 0 (zero), then set it to
the desired value.

- Added a new check to avoid the case where the user
sets the same amount of VFs already configured in
the driver. A new helper function called
'_mlx5_SRIOV_get_numvf_config_file' was created to
avoid code repetition between this new check and
the actual SR-IOV setup.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>