SR-IOV for ConnectX-4: API change
Former API, 'actions', violated the REST best practices because
the method executed upon the resource was changing due to the
value of the parameter 'name'. REST requires that each action
name must clearly state what it is doing with the resource and
the parameters are just the modifiers of that same action.

The solution is to rename it to 'enable_sriov' accepting a single
parameter named 'num_vfs'. If this API is executed in a non-eligible
interface (any interface not loaded with the mlx5_core module)
a 400 Bad Request error will be thrown.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>