Remove any cherrypy configuration from gingerbase.conf file
Ginger configuration file should only contain data that can be
edited by user, which is not the case of the cherrypy configuration.

Let user changes cherrypy configuration may break the Ginger APIs.
To avoid any confusion and problems on that side, move the cherrypy
configuration to Ginger.get_custom_conf().

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>