Issue #304. Check for carrier value of each interface after activation
As a part of the fix a check has been added in activate function
wherein the carrier value for the interface will be checked after
activation and an error message will be thrown if the device is
activated sucessfully but the physical cable of the device is not
connected. Unit tests have been modified for the same.

Signed-off-by: Megha Smriti <mesmriti@linux.vnet.ibm.com>