SR-IOV fixes after hardware testing - nw_interfaces_utils.py
Changes in nw_interfaces_utils:

- Do not throw an error when the OPENIBD_POST_START isn't found
in the /etc/infiniband/openib.conf file. The absence of this
variable was being interpreted as the absence of SR-IOV support,
but real testing showed that this wasn't true. Instead of
throwing an error, simply append the line at the end of the
openib.conf file.

- The SR-IOV setup in boot time works similar as in regular
runtime. That means we need to reset the VFs of the driver
before setting the new value. An 'echo 0 > /sys/(...)' line
was added for each interface entry.

- Ginger backend now will always set the OPENIBD_POST_START
variable under the /etc/infiniband/openib.conf file, even if
it's not the first time the boot script is created.

- Fixed an issue with the boot script generation when updating
an existing entry in the script.

One i18n.py message was modified to reflect the first change
mentioned in this commit message.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>