model/interfaces.py: mlx5_core driver SR-IOV delay workaround
This patch adds a workaround in the interfaces backend
to deal with a strange behavior of the mlx5_core
driver. When setting the virtual functions, the mlx5_core driver
creates each VF as 'eth0' and then rename it to the actual VF
interface name. This is triggered by a write (or echo) call
in a SR-IOV system file. This is a blocker call that takes
several seconds to complete, and the initial assumption was
that the process was completed after this write call is
finished.

However, the driver doesn't finish this procedure in that
period of time. Looking at dmesg, we can see that at least
one of the VFs are being renamed after the 'SRIOV enabled'
message:

[ 1215.488159] mlx5_core 0001:01:00.1 enP1p1s0f1: Link up.
[ 1215.511866] mlx5_core 0001:01:04.1 enP1p1s4f1: Link up.
( ... )
[ 1215.516996] mlx5_core 0001:01:05.1 enP1p1s5f1: Link up.
[ 1215.518166] mlx5_core 0001:01:05.2 eth0: Link up.
[ 1215.518764] (0001:01:00.1): E-Switch: SRIOV enabled: active vports(11)
[ 1215.781648] mlx5_core 0001:01:05.2 enP1p1s5f2: renamed from eth0
[ 1215.805793] systemd-udevd[85515]: renamed network interface eth0 to enP1p1s5f2

The result was that the interface listing would have the
'eth0' interface which would be renamed later on to
'enP1p1s5f2'. The interface left to be renamed is random, it's not
always the same.

Three scenarios would occur in this case:

- If the driver renamed the 'eth0' temp interface before the backend
fetches the interfaces listing, everything would occur as expected.
Testing shows that this scenario is less likely to occur when setting
more than 6 VFs, almost never happening when setting more than
14 VFs.

- Ginger backend get the 'eth0' in the interface listing and
the mlx5_driver driver renames it right before Ginger tries
to fetch its mac address. This is the most common scenario in the
tests made. Ginger aborts the creation of the config file
in this case.

- Ginger backend get the 'eth0' interface and was quickly
enough to get the mac address of the interface using the
'eth0' name. This is the worst scenario and it was happening
too often. The result is that a config file was created linking
the 'eth0' device to the mac address of a VF that would
be renamed to something else ('enP1p1s5f2' for example). When
executing ifdown/ifup or restarting the whole network, the
network script assigns 'eth0' to the mac address of the VF,
renaming the VF back to 'eth0'.

The solution was to introduce a controlled delay after the
SR-IOV setup. Ginger will fetch the newly created interfaces
from time to time, checking if there is an 'eth0' interface
in the listing. If no 'eth0' present, log the time taken
and proceed. A maximum of 5 seconds of delay was set. No
error will be thrown if the time surpass 5 seconds, but
the error log will inform the user that the driver is
taking too long to finish its SR-IOV setup.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>