model/services.py: adding docstrings and pylint fixes
Pylint is one of the most used tools to check for
python code standards, one step ahead of what we
already do with pyflakes. It uses PEP8 to check for
missing docstrings, invalid variable names, no
'self' use on class methods and so on.

This patch  makes changes in services.py to comply
with pylint standards and also eliminate wok_log.error()
redundancies.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>