Add 'port_type' to sanadapters API: UI changes
This patch adds a new column named 'Port type' to
store the new API key 'port_type'.

The 'ports_info' column will return the 'N/A' value
when the port type is 'virtual'. Otherwise, using
the previous logic, it would be set to
'Not applicable/Not applicable'.

Some width redimensions were made to better fit
this new column with the existing content.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>