SR-IOV for ConnectX-4: 0 (zero) VF setup changes
In the scenario where the user set num_vf to zero,
there are some steps that aren't required in the
overall SR-IOV setup because, by default, setting
VF to zero is always the first step.

This patch makes such distinction in model/interfaces.py
while also adding a new error message, specific for the
scenario there is no VFs enabled in the driver and the
user executes the SR-IOV setup with VF = 0.