Issue #348. Activation of physical device and BOND/VLAN.
v3: Changed the error message and removed an error messgae from UI.

v2: Seggregated the error messages for ethernet devices and non ethernet devices
     with iface_type as 'unknown'.

v1: The activation of a physical device leads to creation of carrier file
In case of vlan and bond the file is created and the carrier value in the
file checked for displaying the activation message. The error message has
modified since the in case of vlan and bond the file creation might take a
longer time , therefore we need to refresh and try the same operation again.

Signed-off-by: Megha Smriti <mesmriti@linux.vnet.ibm.com>