Github #195 - adding %dir directives in ginger.spec.in
The problem described in github #195 occurs because the
dirs aren't being declared in the spec file. This means that
the package does not 'own' the directory, only the files
inside it.

This can cause issues when renaming dirs, like described in the
bug. The 'models' and 'controls' dir aren't being erased because
the package didn't own these dirs. Even when the dirs are empty
it will not remove them in the upgrade process.

Although there is no easy fix for the previous versions of Ginger
(I've discarded using a scriplet just for that) this change will
prevent the error from happening in the future. A workaround for
this issue, if the user installed Ginger version 2.0.0, is to
remove the package then install the newest one.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>