Github #121: WoK log flooded by hddtemp messages
This patch eliminates the flood of messages like

'Sensors hdd: /dev/sde: AMI Virtual HDisk3: S.M.A.R.T. not available'

from the Sensor Monitor backend. This message indicates
that the hdd does not have S.M.A.R.T (Self-Monitoring,
Analysis, and Reporting Technology) support and should
be expected to happen (and be properly ignored) in the
output of the hddtemp command.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>