Capabilities API: get current feature state
This patch does the following changes in capabilities.py:

- no need to receive an array of model instances. Using a logic
that was inspired by the get_all_model_instances from WoK utils,
Capabilities is now loading every model instance without the need
of receiving them in the constructor.

- evaluate the state of each feature at the lookup() call. Instead
of evaluating it once and store the result, store the method to
retrieve the state at any given time. This allows for interesting
use cases like turning on the tuned service and seeing the Power
Profiles UI showing up without the need to restart WoK.

With these changes in capabilities.py this patch was able to
also remove some clutter from model.py.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>