test_storage_devs.py: import adjustments
This patch fixes test_storage_devs.py imports after the change made
in the previous commit.

Note that the mock namespace is 'wok.plugins.ginger.model.storage_devs'
because the methods were imported from dasd_utils and utils to storage_devs
and use with this namespace in 'get_final_list'.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>