Fix for the size slider in storage devices.
Sliders on Create Partition on Storage Devices and add Swap Devices for the File type did not take the default slider value 1MB,
as it was not inilized and picked slider values only if it's moved. Fixed it to initialize the variable with 1MB that to be
passed to control code. As part of this bugzilla id, replaced the title Add Partition to Create Partition for Storage Devices
to keep the title uniform.

Signed-off-by: Prakash Ramasamy <prakash.ramasamy@in.ibm.com>