Issue #397 - Fedora 24: Network Configuration - Bring VLAN interface up
The problem is related to IPv4 DHCP, when user try to bring vlan interface up
and there is no dhcp working in that interface, an default message is displayed
in the UI and the interface is not started. Changed nw_interfaces_utils.py file
to include a condition to check if when ifup command returns code 4, if so,
raise an OperationFailed.

Signed-off-by: Bianca Carvalho <bianca@linux.vnet.ibm.com>