Bug fix #396: Network Configuration - Add/Delete BOND interface errors
Changed nw_cfginterfaces_utils file to check if there is DEVICE value in slave
interface file. If not, get "NAME" value instead, If can't find both of them,
show a error message letting user know about it. Also changed "GINNET0094E"
message to let the user know physical cable is not connected to the bond interface
if that is the case. Included 'ip link delete' in case carrier file has 0 value for
bond interfaces, so the user can delete it without any errors.

Signed-off-by: Bianca Carvalho <bianca@linux.vnet.ibm.com>